Migrating to Kotlin. UI tests disabled as there is currently a couple issues between Java and Kotlin

+++ app/src/main/java/bj/discogsbrowser/utils/analytics/AnalyticsTracker.java
@@ -19,7 +19,7 @@ public class AnalyticsTracker
         this.tracker = tracker;
     }
 
-    public void send(String screenName, String category, String action, String label, Long value)
+    public void send(String screenName, String category, String action, String label, int value)
     {
         if (!BuildConfig.DEBUG)
         {
